allowing for existing content

+++ tftpbatch.sh
@@ -7,5 +7,9 @@ while IFS= read -r path; do
     dir="$(dirname "$path")"
     printf "GET /catmaid/%s => /opt/VFB/%s\n" "$path" "$dir"
     ! [ -d "$dir" ] && mkdir -p "/opt/VFB/$dir"
-    curl -o "$path" "$server/$path"
+    if [ ! -e "$path" ]; then
+        curl -o "$path" "$server/$path"
+    else
+        echo $path exists
+    fi
 done < "$1"
